+++ client/web/src/components/MapPage/Locations/LocationsList/Card.js
